[Refactor] Changed shunt to use `CanonicalGround`
I was having trouble getting the Shunt concept to use a
net because it violates the rules of the abstraction to
pass in nets like this.

I've modified this to use the CanonicalGround to allow
shunting components to ground. The `Tap` functionality
can handle any other types where we want to connect an
internal part to another net.